OCR/requirements.txt
numpy<2.0
pillow>=10.0.1,<11.0.0
img2pdf>=0.4.8
pymupdf==1.23.6
opencv-python-headless==4.8.0.76
pytesseract==0.3.10
python-multipart>=0.0.5
konlpy==0.6.0
JPype1>=1.5.0
emoji<2.0
kss==3.4.1
sumy==0.9.0
SQLAlchemy>=1.4
sentencepiece>=0.1.99
torch==2.1.0
transformers==4.36.0
fastapi==0.100.0
uvicorn[standard]==0.22.0
ocrmypdf>=16.10.4
